+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.3
+
+- `petit_httpd` new option:
+  - `-force-https`: to automatically redirect to HTTPS when possible.
+
 ## 1.0.2
 
 - Added server name in headers.